pyproject.toml
[project]
name = "report-generator-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fpdf>=1.7.2",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
